(refactor) Updates Vision Follower tests in CPP and Python
