+++ src/kompass_cpp/tests/vision_follower_test.cpp
@@ -1,0 +1,454 @@
+#include "controllers/rgbd_follower.h"
+#include "datatypes/control.h"
+#include "datatypes/path.h"
+#include "datatypes/tracking.h"
+#include "datatypes/trajectory.h"
+#include "test.h"
+#include "utils/logger.h"
+#include "utils/transformation.h"
+#include <memory>
+#include <opencv2/opencv.hpp>
+#define BOOST_TEST_MODULE KOMPASS TESTS
+#include "json_export.h"
+#include <boost/dll/runtime_symbol_info.hpp> // for program_location
+#include <boost/filesystem.hpp>
+#include <boost/test/included/unit_test.hpp>
+#include <cmath>
+#include <vector>
+
+using namespace Kompass;
+
+struct RGBDFollowerTestConfig {
+  bool use_local_frame;
+
+  // Sampling configuration
+  double timeStep;
+  int predictionHorizon;
+  int controlHorizon;
+
+  double distance_tolerance = 0.1;
+  double target_distance = 0.2;
+  float robot_radius = 0.1;
+
+  // Detected Boxes
+  std::vector<Bbox3D> detected_boxes;
+  int num_test_boxes = 3;
+  Eigen::Vector2i ref_point_img{150, 150};
+  float boxes_ori = 0.0f;
+
+  // Robot configuration
+  Control::LinearVelocityControlParams x_params;
+  Control::LinearVelocityControlParams y_params;
+  Control::AngularVelocityControlParams angular_params;
+  Control::ControlLimitsParams controlLimits;
+  Control::ControlType controlType;
+  Kompass::CollisionChecker::ShapeType robotShapeType;
+  std::vector<float> robotDimensions;
+
+  // Depth detector
+  Eigen::Vector2f focal_length = {911.71, 910.288};
+  Eigen::Vector2f principal_point = {643.06, 366.72};
+  Eigen::Vector2f depth_range = {0.001, 5.0}; // 5cm to 5 meters
+  float depth_conv_factor = 1e-3;             // convert from mm to m
+  Eigen::Isometry3f camera_body_tf;
+
+  // Robot pointcloud values (global frame)
+  std::vector<Path::Point> cloud;
+
+  // Robot start state (pose)
+  Path::State robotState;
+
+  // Tracked target with respect to the robot
+  Control::Velocity2D tracked_vel;
+  Control::TrackedPose2D tracked_pose;
+  // RGBDFollower configuration object
+  std::unique_ptr<Control::RGBDFollower> controller;
+
+  // Constructor to initialize the struct
+  RGBDFollowerTestConfig(const bool use_local_frame = true,
+                         const double timeStep = 0.1,
+                         const int predictionHorizon = 20,
+                         const int controlHorizon = 2, const float maxVel = 2.0,
+                         const float maxOmega = 4.0)
+      : use_local_frame(use_local_frame), timeStep(timeStep),
+        predictionHorizon(predictionHorizon), controlHorizon(controlHorizon),
+        x_params(maxVel, 5.0, 10.0), y_params(1, 3, 5),
+        angular_params(3.14, maxOmega, 3.0, 3.0),
+        controlLimits(x_params, y_params, angular_params),
+        controlType(Control::ControlType::DIFFERENTIAL_DRIVE),
+        robotShapeType(Kompass::CollisionChecker::ShapeType::CYLINDER),
+        robotDimensions{robot_radius, 0.4},
+        robotState(-0.8, 0.0, 0.0, 0.0), tracked_vel(0.1, 0.0, 0.1),
+        tracked_pose(0.0, 0.0, 0.0, tracked_vel) {
+    auto config = Control::RGBDFollower::RGBDFollowerConfig();
+    config.setParameter("control_time_step", timeStep);
+    config.setParameter("control_horizon", controlHorizon);
+    config.setParameter("prediction_horizon", predictionHorizon);
+    config.setParameter("use_local_coordinates", use_local_frame);
+    config.setParameter("target_distance", target_distance);
+    config.setParameter("target_orientation", 0.0);
+    config.setParameter("distance_tolerance", distance_tolerance);
+
+    // Body to (FLU-aligned) camera tf from robot of test pictures.
+    // The DepthDetector applies the optical->FLU rotation internally, so this
+    // tf must be the body->camera-link transform, NOT the body->optical frame.
+    auto body_to_link_tf =
+        getTransformation(Eigen::Quaternionf{0.0f, 0.1987f, 0.0f, 0.98f},
+                          Eigen::Vector3f{0.32f, 0.0209f, 0.3f});
+
+    auto link_to_cam_tf =
+        getTransformation(Eigen::Quaternionf{0.01f, -0.00131f, 0.002f, 0.9999f},
+                          Eigen::Vector3f{0.0f, 0.0105f, 0.0f});
+
+    Eigen::Isometry3f body_to_cam_tf = body_to_link_tf * link_to_cam_tf;
+
+    Eigen::Vector3f translation = body_to_cam_tf.translation();
+    Eigen::Quaternionf rotation_quat =
+        Eigen::Quaternionf(body_to_cam_tf.rotation());
+    Eigen::Vector4f rotation = {rotation_quat.w(), rotation_quat.x(),
+                                rotation_quat.y(), rotation_quat.z()};
+
+    controller = std::make_unique<Control::RGBDFollower>(
+        controlType, controlLimits, robotShapeType, robotDimensions,
+        translation, rotation, config);
+
+    // Initialize the detected boxes
+    Bbox3D new_box;
+    new_box.size = {0.5f, 0.5f, 1.0f};
+    detected_boxes.resize(num_test_boxes);
+    for (int i = 0; i < num_test_boxes; ++i) {
+      auto new_box_shift =
+          Eigen::Vector3f({float(0.7 * i), float(0.7 * i), 0.0f});
+      auto img_frame_shift = Eigen::Vector2i{50 * i, 50 * i};
+      new_box.center = new_box_shift;
+      new_box.center_img_frame = img_frame_shift + ref_point_img;
+      new_box.size_img_frame = {25, 25};
+      new_box.timestamp = 0.0f;
+      detected_boxes[i] = new_box;
+    }
+  }
+
+  void moveDetectedBoxes() {
+    // Update the detected boxes using the velocity command
+    Eigen::Vector3f target_ref_vel = {float(tracked_vel.vx() * cos(boxes_ori)),
+                                      float(tracked_vel.vx() * sin(boxes_ori)),
+                                      0.0f};
+    boxes_ori += tracked_vel.omega() * timeStep;
+    for (auto &box : detected_boxes) {
+      box.center += target_ref_vel * timeStep;
+      box.timestamp += timeStep;
+    }
+  };
+
+
+  // Build a ring of points sitting on the tracked target's circular
+  // footprint (radius = 0.5 * max(box.size.x, box.size.y), matching the
+  // inscribed-circle convention used elsewhere). Returned in world frame —
+  // the caller is responsible for any frame conversion.
+  std::vector<Path::Point> targetBodyPoints(int num_points = 24) const {
+    const double r = 0.5 * std::max(detected_boxes.front().size.x(),
+                                    detected_boxes.front().size.y());
+    std::vector<Path::Point> pts;
+    pts.reserve(num_points);
+    for (int i = 0; i < num_points; ++i) {
+      const double a = (2.0 * M_PI * i) / num_points;
+      pts.emplace_back(tracked_pose.x() + r * std::cos(a),
+                       tracked_pose.y() + r * std::sin(a), 0.1);
+    }
+    return pts;
+  }
+
+  bool test_one_cmd_depth(const std::string image_file_path,
+                          const std::vector<Bbox2D> &detections,
+                          const Eigen::Vector2i &clicked_point,
+                          std::vector<Path::Point> cloud) {
+    // robot velocity
+    Control::Velocity2D cmd;
+    // Get image
+    cv::Mat cv_img = cv::imread(image_file_path, cv::IMREAD_UNCHANGED);
+
+    if (cv_img.empty()) {
+      LOG_ERROR("Could not open or find the image");
+    }
+
+    // Create an Eigen matrix of type int from the OpenCV Mat
+    auto depth_image = Eigen::MatrixX<unsigned short>(cv_img.rows, cv_img.cols);
+    for (int i = 0; i < cv_img.rows; ++i) {
+      for (int j = 0; j < cv_img.cols; ++j) {
+        depth_image(i, j) = cv_img.at<unsigned short>(i, j);
+      }
+    }
+
+    controller->setCameraIntrinsics(focal_length.x(), focal_length.y(),
+                                    principal_point.x(), principal_point.y());
+
+    controller->setCurrentState(robotState);
+    auto found_target = controller->setInitialTracking(
+        clicked_point(0), clicked_point(1), depth_image, detections);
+    if (!found_target) {
+      LOG_WARNING("Point not found on image");
+      return false;
+    } else {
+      LOG_INFO("Point found on image ...");
+    }
+
+    auto res = controller->getTrackingCtrl(depth_image, detections, cmd);
+    if (res.isTrajFound) {
+      LOG_INFO("Got control (vx, vy, omega) = (",
+               res.trajectory.velocities.vx[0], ", ",
+               res.trajectory.velocities.vy[0], ", ",
+               res.trajectory.velocities.omega[0], ")");
+    }
+    return res.isTrajFound;
+  }
+
+  double run_test(const int numPointsPerTrajectory, std::string pltFileName) {
+    Control::TrajectorySamples2D samples(2, numPointsPerTrajectory);
+    Control::TrajectoryVelocities2D simulated_velocities(
+        numPointsPerTrajectory);
+    Control::TrajectoryPath robot_path(numPointsPerTrajectory),
+        tracked_path(numPointsPerTrajectory);
+    // Zero-fill the pre-allocated Eigen vectors so any slots that remain
+    // unfilled (early break from the divergence guard, etc.) plot as zeros
+    // instead of uninitialized garbage / NaN.
+    robot_path.x.setZero();
+    robot_path.y.setZero();
+    robot_path.z.setZero();
+    tracked_path.x.setZero();
+    tracked_path.y.setZero();
+    tracked_path.z.setZero();
+    simulated_velocities.vx.setZero();
+    simulated_velocities.vy.setZero();
+    simulated_velocities.omega.setZero();
+    Control::Velocity2D cmd;
+    Control::TrajSearchResult result;
+
+    controller->setCurrentState(robotState);
+
+    // Set the initial tracking boxes in local or global frame
+    std::vector<Bbox3D> initial_boxes;
+    if (use_local_frame) {
+      // Transform boxes to local coordinates for initialization
+      Eigen::Isometry3f world_in_robot_tf =
+          getTransformation(robotState).inverse();
+      Eigen::Matrix3f abs_rotation = world_in_robot_tf.linear().cwiseAbs();
+
+      for (auto &box : detected_boxes) {
+        Bbox3D box_local(box);
+        box_local.center = world_in_robot_tf * box.center;
+        box_local.size = abs_rotation * box.size;
+        initial_boxes.push_back(box_local);
+      }
+    } else {
+      initial_boxes = detected_boxes;
+    }
+
+    controller->setInitialTracking(ref_point_img(0), ref_point_img(1),
+                                   initial_boxes);
+
+    int step = 0;
+    double start_distance =
+        std::sqrt(std::pow(robotState.x - tracked_pose.x(), 2) +
+                  std::pow(robotState.y - tracked_pose.y(), 2));
+    double end_distance = start_distance;
+
+    while (step < numPointsPerTrajectory) {
+      Path::Point point(robotState.x, robotState.y, 0.0);
+      robot_path.add(step, point);
+      tracked_path.add(step, {tracked_pose.x(), tracked_pose.y(), 0.0});
+      controller->setCurrentState(robotState);
+
+      std::vector<Bbox3D> seen_boxes;
+      if (use_local_frame) {
+        // Transform boxes to local coordinates
+        std::vector<Bbox3D> boxes_local_coordinates;
+        Eigen::Isometry3f world_in_robot_tf =
+            getTransformation(robotState).inverse();
+        Eigen::Matrix3f abs_rotation = world_in_robot_tf.linear().cwiseAbs();
+        for (auto &box : detected_boxes) {
+          // Transform the box to the robot frame
+          Bbox3D box_local(box);
+          box_local.center = world_in_robot_tf * box.center;
+          box_local.size = abs_rotation * box.size;
+          boxes_local_coordinates.push_back(box_local);
+        }
+        seen_boxes = boxes_local_coordinates;
+      } else {
+        seen_boxes = detected_boxes;
+      }
+
+      end_distance = std::sqrt(std::pow(robotState.x - tracked_pose.x(), 2) +
+                               std::pow(robotState.y - tracked_pose.y(), 2));
+
+      controller->setCurrentState(robotState);
+
+      result = controller->getTrackingCtrl(seen_boxes, cmd);
+
+      if (result.isTrajFound) {
+        auto velocities = result.trajectory.velocities;
+        int numSteps = 0;
+        for (auto vel : velocities) {
+          if (numSteps >= controlHorizon) {
+            break;
+          }
+          numSteps++;
+          cmd = vel;
+          LOG_DEBUG(BOLD(FBLU("Robot at: {x: ")), KBLU, robotState.x, RST,
+                    BOLD(FBLU(", y: ")), KBLU, robotState.y, RST,
+                    BOLD(FBLU(", yaw: ")), KBLU, robotState.yaw, RST,
+                    BOLD(FBLU("}")));
+
+          LOG_DEBUG(BOLD(FGRN("Found Control: {Vx: ")), KGRN, vel.vx(), RST,
+                    BOLD(FGRN(", Vy: ")), KGRN, vel.vy(), RST,
+                    BOLD(FGRN(", Omega: ")), KGRN, vel.omega(), RST,
+                    BOLD(FGRN("}")));
+
+          robotState.update(vel, timeStep);
+        }
+
+      } else {
+        LOG_ERROR(BOLD(FRED("RGBDFollower Planner failed with robot at: {x: ")),
+                  KRED, robotState.x, RST, BOLD(FRED(", y: ")), KRED,
+                  robotState.y, RST, BOLD(FRED(", yaw: ")), KRED,
+                  robotState.yaw, RST, BOLD(FRED("}")));
+        break;
+      }
+
+      // Advance target and detected boxes the same number of steps as the
+      // robot to keep the simulation in sync.
+      int applied_steps =
+          result.isTrajFound
+              ? std::min(
+                    controlHorizon,
+                    static_cast<int>(result.trajectory.velocities.vx.size()))
+              : 1;
+      for (int s = 0; s < applied_steps; ++s) {
+        tracked_pose.update(timeStep);
+        moveDetectedBoxes();
+      }
+
+      if (end_distance > 3.0 * start_distance) {
+        LOG_ERROR("Robot is moving too far from tracked target");
+        break;
+      }
+
+      step++;
+    }
+    samples.push_back(simulated_velocities, robot_path);
+    samples.push_back(simulated_velocities, tracked_path);
+
+    // Plot the trajectories (Save to json then run python script for plotting)
+    boost::filesystem::path executablePath = boost::dll::program_location();
+    std::string file_location = executablePath.parent_path().string();
+
+    std::string trajectories_filename = file_location + "/" + pltFileName;
+
+    saveTrajectoriesToJson(samples, trajectories_filename + ".json");
+
+    // Save obstacle cloud so the plot can display it
+    std::string obstacles_filename = file_location + "/" + pltFileName + "_obs";
+    {
+      json j_obs;
+      j_obs["points"] = json::array();
+      for (const auto &pt : cloud) {
+        j_obs["points"].push_back(json{{"x", pt.x()}, {"y", pt.y()}});
+      }
+      std::ofstream obs_file(obstacles_filename + ".json");
+      if (obs_file.is_open()) {
+        obs_file << j_obs.dump(4);
+      }
+    }
+
+    std::string command = "python3 " + file_location +
+                          "/trajectory_sampler_plt.py --samples \"" +
+                          trajectories_filename + "\"";
+
+    // Execute the Python script
+    int res = system(command.c_str());
+    if (res != 0)
+      throw std::system_error(res, std::generic_category(),
+                              "Python script failed with error code");
+
+    // Compute final tracking error
+    return std::sqrt(std::pow(robotState.x - tracked_pose.x(), 2) +
+                     std::pow(robotState.y - tracked_pose.y(), 2));
+  }
+};
+
+BOOST_AUTO_TEST_CASE(Test_RGBDFollower_local) {
+  // Create timer
+  Timer time;
+
+  bool use_local_frame = true;
+
+  RGBDFollowerTestConfig testConfig(use_local_frame);
+
+  int numPointsPerTrajectory = 100;
+
+  double end_distance = testConfig.run_test(
+      numPointsPerTrajectory,
+      std::string("vision_follower_with_tracker_local_frame"));
+
+  // target_distance is the edge-to-edge gap between robot and target body,
+  // so the robot's center settles at:
+  //   robot_radius + target_distance + target_radius
+  // target_radius matches RGBDFollower's inscribed-circle convention.
+  const double target_radius =
+      0.5 * std::max(testConfig.detected_boxes.front().size.x(),
+                     testConfig.detected_boxes.front().size.y());
+  double distance_error = end_distance - testConfig.robot_radius -
+                          testConfig.target_distance - target_radius;
+
+  bool test_passed =
+      std::abs(distance_error) < 2.0 * testConfig.distance_tolerance;
+
+  if (test_passed) {
+    LOG_INFO("Tracking finished successfully. End error: ", distance_error,
+             " < tolerance ", 2.0 * testConfig.distance_tolerance);
+  } else {
+    LOG_ERROR("Tracking Failed! End error: ", distance_error, " > tolerance ",
+              2.0 * testConfig.distance_tolerance);
+  }
+
+  BOOST_TEST(test_passed,
+             "RGBDFollower Failed To Track Target in local frame mode");
+}
+
+BOOST_AUTO_TEST_CASE(Test_RGBDFollower_global_frame) {
+  // Create timer
+  Timer time;
+
+  bool use_local_frame = false;
+
+  RGBDFollowerTestConfig testConfig(use_local_frame);
+
+  int numPointsPerTrajectory = 100;
+
+  double end_distance = testConfig.run_test(
+      numPointsPerTrajectory,
+      std::string("vision_follower_with_tracker_global_frame"));
+
+  // target_distance is the edge-to-edge gap between robot and target body,
+  // so the robot's center settles at:
+  //   robot_radius + target_distance + target_radius
+  // target_radius matches RGBDFollower's inscribed-circle convention.
+  const double target_radius =
+      0.5 * std::max(testConfig.detected_boxes.front().size.x(),
+                     testConfig.detected_boxes.front().size.y());
+  double distance_error = end_distance - testConfig.robot_radius -
+                          testConfig.target_distance - target_radius;
+
+  bool test_passed =
+      std::abs(distance_error) < 2.0 * testConfig.distance_tolerance;
+
+  if (test_passed) {
+    LOG_INFO("Tracking finished successfully. End error: ", distance_error,
+             " < tolerance ", testConfig.distance_tolerance);
+  } else {
+    LOG_ERROR("Tracking Failed! End error: ", distance_error, " > tolerance ",
+              testConfig.distance_tolerance);
+  }
+
+  BOOST_TEST(test_passed,
+             "RGBDFollower Failed To Find Control in global frame mode");
+}
